+++ home-helper-api/src/main/kotlin/home/helper/api/persistence/database/mysql/table/DbHelpPointEarnedAchievementRecord.kt
@@ -7,7 +7,7 @@ import java.time.LocalDate
 
 data class DbHelpPointEarnedAchievementRecord(
     var id: Long? = null,
-    var userId: Int? = null,
+    var userId: Long? = null,
     var earnedDate: LocalDate? = null,
-    var earnedPoint: Int? = null
+    var earnedPoint: Long? = null
 )
