+++ home-helper-api/src/main/kotlin/home/helper/api/persistence/database/mysql/table/DbMHelpItemRecord.kt
@@ -6,6 +6,6 @@ package home.helper.api.persistence.database.mysql.table
 data class DbMHelpItemRecord(
     var id: Long? = null,
     var itemName: String? = null,
-    var helpPoint: Int? = null,
+    var helpPoint: Long? = null,
     var memo: String? = null
 )
